refactor(black-box): split the local WebSocket session state out of the open path
The rejection verdict pushed local-websocket-session.ts to the cognitive-load
warn tier. Socket-state translation and message/close bookkeeping move to their
own module; the session file keeps the open and close lifecycle.
