+++ packages/shared-test/black-box-runner/execution/execute-ws-interaction.ts
@@ -1,6 +1,6 @@
 // deno-lint-ignore-file no-explicit-any
 import { executeLocalWsInteraction } from './execute-local-ws-interaction.ts';
-import { rememberWsCloseEvent } from './local-websocket-session.ts';
+import { rememberWsCloseEvent } from './local-websocket-state.ts';
 import {
     executeRemoteWsInteraction,
     shouldExecuteRemoteWsInteraction
